✨ (sideBarFolderButtons/index.tsx): Add usePostUploadFolders query to handle uploading folders and files 🔧 (constants.ts): Add FOLDERS constant to API URLs for folder-related endpoints 🔧 (folders/index.tsx): Add use-post-upload-folders query to handle uploading folders and files
🔧 (use-update-messages.ts): Remove unused file use-update-messages.ts to clean up the project
🔧 (foldersStore.tsx): Remove unused uploadFolder function to improve code maintainability and reduce complexity
🔧 (folders/index.ts): Remove unused uploadFolder function to simplify the codebase and improve readability
